refactor: rename types directory & file
